chore(storybook): split shared main config helpers
